app/config/__init__.py
from dotenv import load_dotenv
import os
load_dotenv()



class EnvDetails: 
    DATABASE_USER = os.getenv("DATABASE_USER")
    DATABASE_PASSWORD = os.getenv("DATABASE_PASSWORD")
    DATABASE_HOST = os.getenv("DATABASE_HOST")
    DATABASE_PORT = os.getenv("DATABASE_PORT")
    DATABASE_NAME = os.getenv("DATABASE_NAME")
    REDIS_URL = os.getenv("REDIS_URL")

