requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
celery[redis]
prometheus-fastapi-instrumentator
python-dotenv
psycopg2-binary